Form view component replaced duplicated work
